requirements.txt
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
decorator==5.1.1
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.12.0
typing_extensions==4.8.0
wcwidth==0.2.8
